Check for GPS/UAV Terminal when checking if unit can timestamp
GPS' and UAV Terminals both have the capability to display the time in the GPS info panel. Logically, this means the player should be able to mark the time on the map using them.
Therefore, it's only logical that the check for whether or not a player can mark the timestamp accounts for these items.